requirement.txt
numpy==2.2.5
pycryptodome==3.22.0